Removed a misplaced type cast
Former-commit-id: 3b980671d3e5214ff203e471baa14e9976425590 [formerly aca60c7575c8629fb2ee44ce74e878769a1b9139 [formerly 42ed8f91e3354fc9194f9869ff80d029ac67c087]]
Former-commit-id: cc34a9f6603c9422f4573b7fb1b04bd56541d667
Former-commit-id: 06d10863e9fa137770f4604cf1fa546b78170d1c

+++ scripts/sct_straighten_spinalcord.py
@@ -107,7 +107,7 @@ def smooth_centerline(fname_centerline, algo_fitting='hanning', type_window='han
     if phys_coordinates:
         sct.printv('.. Computing physical coordinates of centerline/segmentation...', verbose)
         coord_centerline = np.array(zip(x_centerline, y_centerline, z_centerline))
-        phys_coord_centerline = np.asarray(file_image.transfo_pix2phys(coord_centerline), dtype=int)
+        phys_coord_centerline = np.asarray(file_image.transfo_pix2phys(coord_centerline))
         x_centerline = phys_coord_centerline[:, 0]
         y_centerline = phys_coord_centerline[:, 1]
         z_centerline = phys_coord_centerline[:, 2]
